+++ apps/connect/src/components/modal/modals/DebugSettingsModal.tsx
@@ -57,7 +57,8 @@ export const DebugSettingsModal: React.FC<DebugSettingsModalProps> = props => {
 
     const driveTriggers =
         documentDrives.find(drive => drive.header.id === selectedDrive)?.state
-            .local.triggers || [];
+            .local
+            .triggers || [];
 
     const isEmptyURL = driveUrl === '';
     const disableUrlButtons = !selectedDrive || isEmptyURL;