+++ apps/connect/src/store/processors.ts
@@ -1,4 +1,4 @@
-import { atomStore } from '#store';
+import { atomStore } from '@powerhousedao/common';
 import { type DocumentDriveDocument } from 'document-drive';
 import { ProcessorManager } from 'document-drive/processors/processor-manager';
 import { useAtomValue } from 'jotai';
@@ -18,12 +18,14 @@ async function createProcessorManager() {
 }
 
 export const processorManagerAtom = atomWithLazy(createProcessorManager);
+processorManagerAtom.debugLabel = 'processorManagerAtomInConnect';
 
 // blocks rendering until processor manager is initialized
 export const useProcessorManager = () => useAtomValue(processorManagerAtom);
 
 // will return undefined until processor manager is initialized. Does not block rendering.
 const unwrappedProcessorManager = unwrap(processorManagerAtom);
+unwrappedProcessorManager.debugLabel = 'unwrappedProcessorManagerInConnect';
 
 export const useUnwrappedProcessorManager = () =>
     useAtomValue(unwrappedProcessorManager);