+++ packages/common/state/drives.ts
@@ -34,7 +34,7 @@ export function useDriveById(
   const data = loadableDrives.data;
   return {
     state: "hasData",
-    data: data.find((d) => d?.header.id === id),
+    data: data.find((d) => d.header.id === id),
   };
 }
 
@@ -43,7 +43,7 @@ export function useUnwrappedDriveById(
 ): DocumentDriveDocument | undefined {
   const drives = useUnwrappedDrives();
   if (!id) return undefined;
-  return drives?.find((d) => d?.header.id === id);
+  return drives?.find((d) => d.header.id === id);
 }
 
 export function useSelectedDrive() {