+++ packages/document-drive/src/server/base-server.ts
@@ -1674,6 +1674,20 @@ export class BaseDocumentDriveServer
         throw error;
       }
 
+      this.eventEmitter.emit(
+        "documentOperationsAdded",
+        driveId,
+        documentId,
+        operations,
+      );
+
+      this.eventEmitter.emit(
+        "operationsAdded",
+        driveId,
+        documentId,
+        operations,
+      );
+
       return {
         status: "SUCCESS",
         document,
@@ -1963,6 +1977,9 @@ export class BaseDocumentDriveServer
       if (error) {
         throw error;
       }
+
+      this.eventEmitter.emit("driveOperationsAdded", driveId, operations);
+      this.eventEmitter.emit("operationsAdded", driveId, null, operations);
 
       return {
         status: "SUCCESS",