+++ packages/design-system/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 3.2.0-dev.5 (2025-06-26)
+
+### 🚀 Features
+
+- **common:** add atoms library ([dbc8e8b44](https://github.com/powerhouse-inc/powerhouse/commit/dbc8e8b44))
+
+### 🩹 Fixes
+
+- **connect,codegen,common,reactor-browser:** fix analytics query subscription ([6e9729739](https://github.com/powerhouse-inc/powerhouse/commit/6e9729739))
+- **common,design-system,reactor-browser:** fix drag and drop on generic drive explorer ([9b5113e74](https://github.com/powerhouse-inc/powerhouse/commit/9b5113e74))
+
+### ❤️ Thank You
+
+- acaldas
+- ryanwolhuter
+
 ## 3.2.0-dev.4 (2025-06-25)
 
 ### 🚀 Features